app/requirements.txt
django
Pillow
django-bootstrap3
django-crispy-forms
django-dotenv
python-social-auth[django]
django-extensions
psycopg2
social-auth-app-django
django-gulp
